+++ ZenTune/Assets/core/config.py
@@ -7,8 +7,8 @@ from pathlib import Path
 
 from Assets.core.platform import RUNTIME_DIR
 
-LOCAL_VERSION = "2.0.1"
-LOCAL_BUILD = "R2-09Jul26-f1"
+LOCAL_VERSION = "2.0.2"
+LOCAL_BUILD = "R2-11Aug26-s2"
 
 APP_NAME = "zentune"
 DIST_NAME = "ZenTune"